requirements.txt
openai==0.28.0
python-dotenv==1.0.0
Flask==2.3.3
requests==2.31.0
openai-whisper==20230314
ffmpeg==1.4
ffprobe==0.5
numpy<1.23.0
soundfile==0.12.1
pydub==0.25.1